requirements.txt
rangehttpserver==1.3.3
yt_dlp
yt-dlp-ejs
curl-cffi
discord
python-dotenv
gitpython
pillow
aiohttp